version: реализовать работу с БД (ТЗ-11)

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -3,9 +3,9 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.EntityNotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
-import ru.yandex.practicum.filmorate.storage.UserStorage;
 
 import java.util.List;
 
@@ -14,30 +14,30 @@ import java.util.List;
 @Slf4j
 public class FilmService {
     private final FilmStorage filmStorage;
-    private final UserStorage userStorage;
 
-    public void addLike(int id, long userId) {
-        Film film = filmStorage.getFilm(id);
-        film.addLike();
-        userStorage.getUser(userId).getLikedFilms().add(id);
-        log.info("Like has been added");
+    public void addLike(int filmId, long userId) {
+        if (filmId <= 0) {
+            throw new EntityNotFoundException("Invalid Film ID");
+        }
+        if (userId <= 0) {
+            throw new EntityNotFoundException("Invalid User ID");
+        }
+        filmStorage.addLike(filmId, userId);
     }
 
-    public void deleteLike(int id, long userId) {
-        Film film = filmStorage.getFilm(id);
-        film.deleteLike();
-        userStorage.getUser(userId).getLikedFilms().remove(id);
-        log.info("Like has been deleted");
+    public void deleteLike(int filmId, long userId) {
+        if (filmId <= 0) {
+            throw new EntityNotFoundException("Invalid Film ID");
+        }
+        if (userId <= 0) {
+            throw new EntityNotFoundException("Invalid User ID");
+        }
+        filmStorage.deleteLike(filmId, userId);
     }
 
     public List<Film> getPopular(int count) {
         if (count < 1) {
             throw new IllegalArgumentException("Illegal count value");
-        }
-        if (count == 1) {
-            log.info("The most popular film requested");
-        } else {
-            log.info(count + " most popular films requested");
         }
         return filmStorage.getPopular(count);
     }
@@ -47,6 +47,9 @@ public class FilmService {
     }
 
     public Film getFilm(int id) {
+        if (id <= 0) {
+            throw new EntityNotFoundException("Invalid Film ID");
+        }
         return filmStorage.getFilm(id);
     }
 
@@ -59,6 +62,9 @@ public class FilmService {
     }
 
     public void deleteFilm(int id) {
+        if (id <= 0) {
+            throw new EntityNotFoundException("Invalid Film ID");
+        }
         filmStorage.deleteFilm(id);
     }
 }
